+++ src/main/java/step2/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package step2_calculator;
+package step2.calculator;
 
 public class Calculator {
