+++ src/main/java/step3/racingCar/Car.java
@@ -1,4 +1,4 @@
-package step3_racingCar;
+package step3.racingCar;
 /*
 * 자동차 경주의 자동차를 담당하는 클래스
 *
