+++ src/main/java/step3/racingCar/InputView.java
@@ -1,4 +1,4 @@
-package step3_racingCar;
+package step3.racingCar;
 
 import java.util.ArrayList;
 import java.util.Scanner;
