+++ src/main/java/step3/racingCar/ResultView.java
@@ -1,4 +1,4 @@
-package step3_racingCar;
+package step3.racingCar;
 
 import java.util.LinkedList;
