+++ src/test/java/step2/calculator/CalculatorTest.java
@@ -1,11 +1,9 @@
-package step2_calculator;
+package step2.calculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import step2_calculator.Calculator;
-import step2_calculator.Util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
